fix: save and check installed status of plugins
this is for plugins that need a post install step.
miq will mark and check if it has been "fully installed"
by whether the post install ran successfully. if it
hasnt then it will permit installing

(maybe i should just make it run the post install instead of
attempting to clone again...)